# refactor(sumologicextension): use bytes slices and strings.Builder to decrease allocations