```
$ benchcmp main.txt pr.txt
benchmark                                          old ns/op     new ns/op     delta
Benchmark_PrometheusFormatter_Metric2String-16     66901         53862         -19.49%
Benchmark_PrometheusFormatter_Metric2String-16     65607         53773         -18.04%
Benchmark_PrometheusFormatter_Metric2String-16     65205         54047         -17.11%

benchmark                                          old allocs     new allocs     delta
Benchmark_PrometheusFormatter_Metric2String-16     696            512            -26.44%
Benchmark_PrometheusFormatter_Metric2String-16     696            512            -26.44%
Benchmark_PrometheusFormatter_Metric2String-16     696            512            -26.44%

benchmark                                          old bytes     new bytes     delta
Benchmark_PrometheusFormatter_Metric2String-16     20326         15793         -22.30%
Benchmark_PrometheusFormatter_Metric2String-16     20341         15785         -22.40%
Benchmark_PrometheusFormatter_Metric2String-16     20315         15811         -22.17%
```