@pmalek-sumo 2022-04-05
> Have you benchmarked how much of the improvement comes from introducing StringBuilder vs the other changes? The StringBuilder code is much uglier, and intuitively it shouldn't be that much faster than fmt.Sprintf for short strings.

Just those 2 changes introducing `strings.Builder` in `tags2String()` contribute ~~20-25% of allocs and similarly in time:

```
benchmark                                          old ns/op     new ns/op     delta
Benchmark_PrometheusFormatter_Metric2String-16     49478         59523         +20.30%
Benchmark_PrometheusFormatter_Metric2String-16     50142         64352         +28.34%
Benchmark_PrometheusFormatter_Metric2String-16     49223         59849         +21.59%

benchmark                                          old allocs     new allocs     delta
Benchmark_PrometheusFormatter_Metric2String-16     538            674            +25.28%
Benchmark_PrometheusFormatter_Metric2String-16     538            674            +25.28%
Benchmark_PrometheusFormatter_Metric2String-16     538            674            +25.28%

benchmark                                          old bytes     new bytes     delta
Benchmark_PrometheusFormatter_Metric2String-16     16950         19635         +15.84%
Benchmark_PrometheusFormatter_Metric2String-16     16923         19645         +16.08%
Benchmark_PrometheusFormatter_Metric2String-16     16938         19654         +16.03%
```