@pmalek-sumo 2022-04-06
Another comparison:

```go
	returnValue := make([]string, 0, length)
	mergedAttributes.Range(func(k string, v pdata.AttributeValue) bool {
		key := f.sanitizeKeyBytes([]byte(k))
		value := f.sanitizeValue(v.AsString())

		returnValue = append(
			returnValue,
			fmt.Sprintf(
				`%s="%s"`,
				key, value,
			),
		)
		return true
	})

	return prometheusTags(fmt.Sprintf("{%s}", strings.Join(returnValue, ",")))
```

```
goos: darwin
goarch: amd64
pkg: github.com/SumoLogic/sumologic-otel-collector/pkg/exporter/sumologicexporter
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
Benchmark_PrometheusFormatter_Metric2String
Benchmark_PrometheusFormatter_Metric2String-16             20386             58591 ns/op           19630 B/op        674 allocs/op
Benchmark_PrometheusFormatter_Metric2String-16             20182             59102 ns/op           19668 B/op        674 allocs/op
Benchmark_PrometheusFormatter_Metric2String-16             19946             61476 ns/op           19647 B/op        674 allocs/op
PASS
ok      github.com/SumoLogic/sumologic-otel-collector/pkg/exporter/sumologicexporter    5.821s
```

---

using `strings.Builder`

```go
	returnValue := make([]string, 0, length)
	mergedAttributes.Range(func(k string, v pdata.AttributeValue) bool {
		key := f.sanitizeKeyBytes([]byte(k))
		value := f.sanitizeValue(v.AsString())

		sb := strings.Builder{}
		sb.Grow(len(key) + len(value) + 1 + 2)
		sb.Write(key)
		sb.WriteRune('=')
		sb.WriteRune('"')
		sb.WriteString(value)
		sb.WriteRune('"')

		returnValue = append(
			returnValue,
			fmt.Sprintf(
				"%s", sb.String(),
			),
		)
		return true
	})

	ret := strings.Join(returnValue, ",")
	sb := strings.Builder{}
	sb.Grow(2 + len(ret))
	sb.WriteRune('{')
	sb.WriteString(ret)
	sb.WriteRune('}')

	return prometheusTags(sb.String())
```

```
goos: darwin
goarch: amd64
pkg: github.com/SumoLogic/sumologic-otel-collector/pkg/exporter/sumologicexporter
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
Benchmark_PrometheusFormatter_Metric2String
Benchmark_PrometheusFormatter_Metric2String-16             21086             55722 ns/op           18930 B/op        658 allocs/op
Benchmark_PrometheusFormatter_Metric2String-16             21694             57132 ns/op           18978 B/op        658 allocs/op
Benchmark_PrometheusFormatter_Metric2String-16             21644             58168 ns/op           18953 B/op        658 allocs/op
PASS
ok      github.com/SumoLogic/sumologic-otel-collector/pkg/exporter/sumologicexporter    5.771s
```

---

using `strings.Builder` and no `fmt`

```go
	returnValue := make([]string, 0, length)
	mergedAttributes.Range(func(k string, v pdata.AttributeValue) bool {
		key := f.sanitizeKeyBytes([]byte(k))
		value := f.sanitizeValue(v.AsString())

		sb := strings.Builder{}
		sb.Grow(len(key) + len(value) + 1 + 2)
		sb.Write(key)
		sb.WriteRune('=')
		sb.WriteRune('"')
		sb.WriteString(value)
		sb.WriteRune('"')

		returnValue = append(
			returnValue,
			sb.String(),
		)
		return true
	})

	ret := strings.Join(returnValue, ",")
	sb := strings.Builder{}
	sb.Grow(2 + len(ret))
	sb.WriteRune('{')
	sb.WriteString(ret)
	sb.WriteRune('}')

	return prometheusTags(sb.String())
```

```
goos: darwin
goarch: amd64
pkg: github.com/SumoLogic/sumologic-otel-collector/pkg/exporter/sumologicexporter
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
Benchmark_PrometheusFormatter_Metric2String
Benchmark_PrometheusFormatter_Metric2String-16             24160             51110 ns/op           16954 B/op        538 allocs/op
Benchmark_PrometheusFormatter_Metric2String-16             23751             50131 ns/op           16942 B/op        538 allocs/op
Benchmark_PrometheusFormatter_Metric2String-16             24018             49481 ns/op           16944 B/op        538 allocs/op
PASS
ok      github.com/SumoLogic/sumologic-otel-collector/pkg/exporter/sumologicexporter    5.389s
```