@swiatekm-sumo 2022-04-06
Pretty significant difference after all, then. Ok, I'm in favor of using the StringBuilders, let's just make it clear why we're doing it this way with comments and possibly moving the key=value formatting to a separate function.