@pmalek-sumo 2022-04-06
> LGTM. Do you reckon we should add separate unit tests for `stringsJoinAndSurround` or are we sufficiently covered by existing ones?

We could in theory. OTOH it's tested indirectly in other tests.